Get rid of public testing code.
